test: Payment 결제 Test 코드 작성
